# Manually set alpha of all pixels rather than using ImageSharp's clear option