--- a/osu.Framework/IO/Stores/GlyphStore.cs
+++ b/osu.Framework/IO/Stores/GlyphStore.cs
@@ -128,7 +128,7 @@ namespace osu.Framework.IO.Stores
             var page = GetPageImageForCharacter(character);
             LoadedGlyphCount++;
 
-            var image = new Image<Rgba32>(SixLabors.ImageSharp.Configuration.Default, character.Width, character.Height, new Rgba32(255, 255, 255, 0));
+            var image = new Image<Rgba32>(SixLabors.ImageSharp.Configuration.Default, character.Width, character.Height);
 
             var dest = image.GetPixelSpan();
             var source = page.GetPixelSpan();
@@ -137,13 +137,13 @@ namespace osu.Framework.IO.Stores
             int readableHeight = Math.Min(character.Height, page.Height - character.Y);
             int readableWidth = Math.Min(character.Width, page.Width - character.X);
 
-            for (int y = 0; y < readableHeight; y++)
+            for (int y = 0; y < character.Height; y++)
             {
                 int readOffset = (character.Y + y) * page.Width + character.X;
                 int writeOffset = y * character.Width;
 
-                for (int x = 0; x < readableWidth; x++)
-                    dest[writeOffset + x] = source[readOffset + x];
+                for (int x = 0; x < character.Width; x++)
+                    dest[writeOffset + x] = x < readableWidth && y < readableHeight ? source[readOffset + x] : new Rgba32(1, 1, 1, 0);
             }
 
             return new TextureUpload(image);