- [x] Depends on #3037.

Before:
```
|                   Method | FetchCount |      Mean |    Error |    StdDev |    Median |      Gen 0 |      Gen 1 |      Gen 2 |    Allocated |
|------------------------- |----------- |----------:|---------:|----------:|----------:|-----------:|-----------:|-----------:|-------------:|
| BenchmarkRawCachingReuse |        100 |  23.14 ms | 0.461 ms |  1.165 ms |  22.69 ms |   187.5000 |    31.2500 |          - |    995.12 KB |
|         BenchmarkNoCache |        100 | 488.37 ms | 9.755 ms | 23.183 ms | 485.54 ms | 54000.0000 | 53000.0000 | 48000.0000 | 422745.17 KB |
```

After:
```
|                   Method | FetchCount |      Mean |    Error |   StdDev |      Gen 0 |      Gen 1 |      Gen 2 |    Allocated |
|------------------------- |----------- |----------:|---------:|---------:|-----------:|-----------:|-----------:|-------------:|
| BenchmarkRawCachingReuse |        100 |  22.37 ms | 0.445 ms | 0.512 ms |   187.5000 |    62.5000 |          - |    995.13 KB |
|         BenchmarkNoCache |        100 | 470.67 ms | 8.917 ms | 9.157 ms | 55000.0000 | 53000.0000 | 48000.0000 | 422805.16 KB |
```

Minimal saving, but worth it as we were already rewriting most every pixel.